Use a single spawn idle
